myproject/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-restframework==0.0.1
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
Markdown==3.8.2
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
